change the display of picked calories, change dto MBR from int to double to return proper value from api request

+++ app/src/main/java/com/example/fitnessapp/profile_feature/data/remote/dto/Data.kt
@@ -1,6 +1,6 @@
 package com.example.fitnessapp.profile_feature.data.remote.dto
 
 data class Data(
-    val BMR: Int,
+    val BMR: Double,
     val goals: Goals
 )
